UBUNTU: [Config] Change CONFIG_LEDS_TRIGGER_HEARTBEAT from module to built-in in Omap
Most Omap based boards use the heartbeat trigger as one of their main
leds (like beagleboard). The trigger architecture doesn't automatically
load the board's default trigger modules, so when we compile heartbeat
as module it'll only be used when the user manually loads it.
As it'd be good to have it properly loaded since the boot start,
changing it to built-in will make the led to work right after loading
the kernel.

Having a hartbeat LED is very useful during development, as we can
easily notice if the board is still working after a temporary hang, so
using it as default sounds a good idea.

Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>