(pre-stable) mm: Move vma_stack_continue into mm.h
So it can be used by all that need to check for that.

Impact: When introducing the stack guard page, a follow up patch tried to
minimize the effects user-space sees from this change. Unfortunately a
lot of the checks did not take into account that user-space may mlock a
portion of the stack, but not all of it. This causes the stack vma to
be split and a lot of assumptions to go down the drain. Most places have
been fixed by now but looking at /proc/<pid>/maps when a stack vma has
been split still hides a page on every area and not only on head of it
as it was intended.

Fix: The following upstream change makes one inline available to other
code and uses that when checking for the guard page in the proc output.

Testcase: A program that mlocks an area of the stack and then dumps its
maps will see holes before the patch and none after.

Note: When providing incremental patches to Linus, the cc for stable got
lost, though I pinged Greg, but have not received an answer yet. In the
end it hopefully comes from stable. I have not yet created a bug report
for it as I am not really sure how urgent this is to fix.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry-picked from commit 39aa3cb3e8250db9188a6f1e3fb62ffa1a717678 upstream)
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>