UBUNTU: SAUCE: ALSA: HDA: Enable internal mic on Dell E6410 and Dell E6510
BugLink: http://launchpad.net/bugs/628961
BugLink: http://launchpad.net/bugs/605047

This patch will enable internal mic on Dell E6410 and Dell E6510. It
will not go upstream since Takashi decided on a different solution.
Takashi's solution is a bit invasive though, so it won't land for
Maverick, so we'll need this less invasive one instead.  We should be
able to drop this patch in Natty as it should be obsoleted by Takashi's
solution.

Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>