UBUNTU: SAUCE: hid: ntrig: New ghost-filtering event logic
The ntrig devices are notoriously poor at resolving ghost touches,
which affects almost every device on the market. Moreoever, the
performance degrades over time, rendering the screen unusable.
In order to work out of the box, the driver needs to cater for this
deficiency, and should ideally do so without any special tuning.

This patch attempts to paper over this problem, by detecting and
filtering out ghost touches using additional knowledge of the sensor
mechanism. A small number of dimensionless parameters are used, to
keep the complexity at a minimum. The driver has been reported to
work well on Dell Studio 17, Dell XT2, HP TX2 and Lenovo T410s.

Also added Canonical, Ltd. to the copyright.

Not likely to go upstream in its present form. However, it does help
constitue a tuning-free driver which works well together with Unity,
and as such are vital to the MT push.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
Acked-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>