aufs: bugfix, race around Magic SysRq, s_inodes and s_files
Protect sb->s_inodes and ->s_files from aufs Magic SysRq handler.
For allocating inodes, VFS calls s_op->alloc_inode() and then adds it to
s_inodes. Also for destroying, VFS removes it from s_inodes first and
then calls s_op->destory_inode().
With this sequence, it is guranteed that the inode in s_inodes is always
correct and we can refer struct au_iinfo.

But for files, VFS adds the file object to s_files before calling
f_op->open(), and calls ->release() before removing it from s_files.
It means the file in s_files MAY be incorrect and we may not refer
struct au_finfo.
To support the file which is not initialized yet, aufs tests
file->private_data. And to address the file which is released but still
in s_files, aufs removes it from s_files first in file release operation.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>