UBUNTU: execute module-inclusion within a subshell
BugLink: http://bugs.launchpad.net/bugs/621175

When uploading the kernel with the orig.tar.gz included, executable
modes are not represented in the diff, eg:

  dpkg-source: warning: executable mode 0755 of 'debian/scripts/module-inclusion' will not be represented in diff

As a result, when debian/scripts/module-inclusion is called when
building the -virtual kernel .deb, we see the following error:

  /bin/bash: line 1: debian/scripts/module-inclusion: Permission denied

This results in a larger -virtual kernel .deb size because the
module-inclusion script is not run which prevents the removal of modules
not in the inclusion list.  The following patch calls
debian/scripts/module-inclusion from within a subshell in order for it
to execute.

Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>