UBUNTU: SAUCE: Adding vdd_sdi regulator supply to OMAP3EVM
https://bugs.launchpad.net/bugs/630885

When booting, the omapdss subsystem is looking for a regulator named
"vdds_sdi". When the regulator is not found the initialisation sequence
is aborted resulting in omapfb not finding a display to work with.  This
patch allows the omapfb sub system to complete its initialisation
properly and enable LCD display. The problem was fixed by lumping a
"vdds_sdi" with the already existing "vdds_dsi" regulator. This fix
takes its root from work done on the Beagle board and the Pandora board.
More spefically:

 - 7b097896e4a5b5ea4798db806e63a1138b1b8eb8
 - f6873eedd94df20cfb705856a78440a4c176c6b6

Signed-off-by: Mathieu Poirier <mathieu.poirier@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>